move all json codecs under JsonCodecs trait;
