+++ src/main/scala/org/ergoplatform/JsonCodecs.scala
@@ -5,15 +5,18 @@ import java.math.BigInteger
 import cats.syntax.either._
 import io.circe._
 import io.circe.syntax._
-import org.ergoplatform.ErgoBox.{NonMandatoryRegisterId, TokenId}
+import org.ergoplatform.ErgoBox.{BoxId, NonMandatoryRegisterId, TokenId}
 import org.ergoplatform.settings.ErgoAlgos
+import org.ergoplatform.validation.{SigmaValidationSettings, SigmaValidationSettingsSerializer}
 import scorex.crypto.authds.{ADDigest, ADKey}
 import scorex.crypto.hash.Digest32
 import scorex.util.ModifierId
 import sigmastate.Values.{ErgoTree, EvaluatedValue}
+import sigmastate.eval.Extensions._
 import sigmastate.eval.{CGroupElement, CPreHeader, WrapperOf, _}
+import sigmastate.interpreter.{ContextExtension, ProverResult}
 import sigmastate.serialization.{ErgoTreeSerializer, ValueSerializer}
-import sigmastate.{AvlTreeData, SType}
+import sigmastate.{AvlTreeData, AvlTreeFlags, SType}
 import special.collection.Coll
 import special.sigma.{Header, PreHeader}
 
@@ -66,6 +69,20 @@ trait JsonCodecs {
 
   implicit val digest32Encoder: Encoder[Digest32] = _.array.asJson
   implicit val digest32Decoder: Decoder[Digest32] = bytesDecoder(Digest32 @@ _)
+
+  implicit val assetEncoder: Encoder[(TokenId, Long)] = { asset =>
+    Json.obj(
+      "tokenId" -> asset._1.asJson,
+      "amount" -> asset._2.asJson
+    )
+  }
+
+  implicit val assetDecoder: Decoder[(TokenId, Long)] = { cursor =>
+    for {
+      tokenId <- cursor.downField("tokenId").as[TokenId]
+      amount <- cursor.downField("amount").as[Long]
+    } yield (tokenId, amount)
+  }
 
   implicit val modifierIdEncoder: Encoder[ModifierId] = _.asInstanceOf[String].asJson
   implicit val modifierIdDecoder: Decoder[ModifierId] = ModifierId @@ _.as[String]
@@ -160,6 +177,78 @@ trait JsonCodecs {
     }
   }
 
+  implicit val dataInputEncoder: Encoder[DataInput] = { input =>
+    Json.obj(
+      "boxId" -> input.boxId.asJson,
+    )
+  }
+
+  implicit val dataInputDecoder: Decoder[DataInput] = { cursor =>
+    for {
+      boxId <- cursor.downField("boxId").as[ADKey]
+    } yield DataInput(boxId)
+  }
+
+  implicit val inputEncoder: Encoder[Input] = { input =>
+    Json.obj(
+      "boxId" -> input.boxId.asJson,
+      "spendingProof" -> input.spendingProof.asJson
+    )
+  }
+
+  implicit val inputDecoder: Decoder[Input] = { cursor =>
+    for {
+      boxId <- cursor.downField("boxId").as[ADKey]
+      proof <- cursor.downField("spendingProof").as[ProverResult]
+    } yield Input(boxId, proof)
+  }
+
+
+  implicit val contextExtensionEncoder: Encoder[ContextExtension] = { extension =>
+    extension.values.map { case (key, value) =>
+      key -> evaluatedValueEncoder(value)
+    }.asJson
+  }
+
+  implicit val contextExtensionDecoder: Decoder[ContextExtension] = { cursor =>
+    for {
+      values <- cursor.as[Map[Byte, EvaluatedValue[SType]]]
+    } yield ContextExtension(values)
+  }
+
+  implicit val proverResultEncoder: Encoder[ProverResult] = { v =>
+    Json.obj(
+      "proofBytes" -> v.proof.asJson,
+      "extension" -> v.extension.asJson
+    )
+  }
+
+  implicit val proverResultDecoder: Decoder[ProverResult] = { cursor =>
+    for {
+      proofBytes <- cursor.downField("proofBytes").as[Array[Byte]]
+      extMap <- cursor.downField("extension").as[Map[Byte, EvaluatedValue[SType]]]
+    } yield ProverResult(proofBytes, ContextExtension(extMap))
+  }
+
+
+  implicit val avlTreeDataEncoder: Encoder[AvlTreeData] = { v =>
+    Json.obj(
+      "digest" -> v.digest.asJson,
+      "treeFlags" -> v.treeFlags.serializeToByte.asJson,
+      "keyLength" -> v.keyLength.asJson,
+      "valueLength" -> v.valueLengthOpt.asJson
+    )
+  }
+
+  implicit val avlTreeDataDecoder: Decoder[AvlTreeData] = { cursor =>
+    for {
+      digest <- cursor.downField("digest").as[ADDigest]
+      treeFlagsByte <- cursor.downField("treeFlags").as[Byte]
+      keyLength <- cursor.downField("keyLength").as[Int]
+      valueLength <- cursor.downField("valueLength").as[Option[Int]]
+    } yield new AvlTreeData(digest, AvlTreeFlags(treeFlagsByte), keyLength, valueLength)
+  }
+
   implicit val ergoTreeEncoder: Encoder[ErgoTree] = { value =>
     ErgoTreeSerializer.DefaultSerializer.serializeErgoTree(value).asJson
   }
@@ -172,5 +261,111 @@ trait JsonCodecs {
 
   implicit val ergoTreeDecoder: Decoder[ErgoTree] = {
     decodeErgoTree(_.asInstanceOf[ErgoTree])
+  }
+
+  implicit val ergoBoxEncoder: Encoder[ErgoBox] = { box =>
+    Json.obj(
+      "boxId" -> box.id.asJson,
+      "value" -> box.value.asJson,
+      "ergoTree" -> ErgoTreeSerializer.DefaultSerializer.serializeErgoTree(box.ergoTree).asJson,
+      "assets" -> box.additionalTokens.toArray.toSeq
+        .asJson, //(Encoder.encodeSeq(Encoder.encodeTuple2(digest32Encoder, Encoder.encodeLong))),
+      "creationHeight" -> box.creationHeight.asJson,
+      "additionalRegisters" -> box.additionalRegisters.asInstanceOf[Map[NonMandatoryRegisterId, EvaluatedValue[SType]]].asJson,
+      "transactionId" -> box.transactionId.asJson,
+      "index" -> box.index.asJson
+    )
+  }
+
+  implicit val ergoBoxDecoder: Decoder[ErgoBox] = { cursor =>
+    for {
+      value <- cursor.downField("value").as[Long]
+      ergoTreeBytes <- cursor.downField("ergoTree").as[Array[Byte]]
+      additionalTokens <- cursor.downField("assets").as(Decoder.decodeSeq(assetDecoder))
+      creationHeight <- cursor.downField("creationHeight").as[Int]
+      additionalRegisters <- cursor.downField("additionalRegisters").as[Map[NonMandatoryRegisterId, EvaluatedValue[SType]]]
+      transactionId <- cursor.downField("transactionId").as[ModifierId]
+      index <- cursor.downField("index").as[Short]
+    } yield new ErgoBox(
+      value = value,
+      ergoTree = ErgoTreeSerializer.DefaultSerializer.deserializeErgoTree(ergoTreeBytes),
+      additionalTokens = additionalTokens.toColl,
+      additionalRegisters = additionalRegisters,
+      transactionId = transactionId,
+      index = index,
+      creationHeight = creationHeight
+    )
+  }
+
+  implicit val ergoLikeTransactionEncoder: Encoder[ErgoLikeTransaction] = { tx =>
+    Json.obj(
+      "id" -> tx.id.asJson,
+      "inputs" -> tx.inputs.asJson,
+      "dataInputs" -> tx.dataInputs.asJson,
+      "outputs" -> tx.outputs.asJson
+    )
+  }
+
+  implicit val transactionOutputsDecoder: Decoder[(ErgoBoxCandidate, Option[BoxId])] = { cursor =>
+    for {
+      maybeId <- cursor.downField("boxId").as[Option[BoxId]]
+      value <- cursor.downField("value").as[Long]
+      creationHeight <- cursor.downField("creationHeight").as[Int]
+      ergoTree <- cursor.downField("ergoTree").as[ErgoTree]
+      assets <- cursor.downField("assets").as[Seq[(ErgoBox.TokenId, Long)]] // TODO optimize: encode directly into Coll avoiding allocation of Tuple2 for each element
+      registers <- cursor.downField("additionalRegisters").as[Map[NonMandatoryRegisterId, EvaluatedValue[SType]]]
+    } yield (new ErgoBoxCandidate(value, ergoTree, creationHeight, assets.toColl, registers), maybeId)
+  }
+
+  implicit val ergoLikeTransactionDecoder: Decoder[ErgoLikeTransaction] = { implicit cursor =>
+    for {
+      inputs <- cursor.downField("inputs").as[IndexedSeq[Input]]
+      dataInputs <- cursor.downField("dataInputs").as[IndexedSeq[DataInput]]
+      outputsWithIndex <- cursor.downField("outputs").as[IndexedSeq[(ErgoBoxCandidate, Option[BoxId])]]
+    } yield new ErgoLikeTransaction(inputs, dataInputs, outputsWithIndex.map(_._1))
+  }
+
+  implicit val sigmaValidationSettingsEncoder: Encoder[SigmaValidationSettings] = { v =>
+    SigmaValidationSettingsSerializer.toBytes(v).asJson
+  }
+
+  implicit val sigmaValidationSettingsDecoder: Decoder[SigmaValidationSettings] = { implicit cursor: ACursor =>
+    cursor.as[Array[Byte]] flatMap { bytes =>
+      fromThrows(SigmaValidationSettingsSerializer.fromBytes(bytes))
+    }
+  }
+
+  implicit val ergoLikeContextEncoder: Encoder[ErgoLikeContext] = { ctx =>
+    Json.obj(
+      "lastBlockUtxoRoot" -> ctx.lastBlockUtxoRoot.asJson,
+      "headers" -> ctx.headers.toArray.toSeq.asJson,
+      "preHeader" -> ctx.preHeader.asJson,
+      "dataBoxes" -> ctx.dataBoxes.asJson,
+      "boxesToSpend" -> ctx.boxesToSpend.asJson,
+      // TODO: handle unsigned tx
+      "spendingTransaction" -> ctx.spendingTransaction.asInstanceOf[ErgoLikeTransaction].asJson,
+      "selfIndex" -> ctx.selfIndex.asJson,
+      "extension" -> ctx.extension.asJson,
+      "validationSettings" -> ctx.validationSettings.asJson,
+      "costLimit" -> ctx.costLimit.asJson,
+      "initCost" -> ctx.initCost.asJson
+    )
+  }
+
+  implicit val ergoLikeContextDecoder: Decoder[ErgoLikeContext] = { cursor =>
+    for {
+      lastBlockUtxoRoot <- cursor.downField("lastBlockUtxoRoot").as[AvlTreeData]
+      headers <- cursor.downField("headers").as[Seq[Header]]
+      preHeader <- cursor.downField("preHeader").as[PreHeader]
+      dataBoxes <- cursor.downField("dataBoxes").as[IndexedSeq[ErgoBox]]
+      boxesToSpend <- cursor.downField("boxesToSpend").as[IndexedSeq[ErgoBox]]
+      spendingTransaction <- cursor.downField("spendingTransaction").as[ErgoLikeTransaction]
+      selfIndex <- cursor.downField("selfIndex").as[Int]
+      extension <- cursor.downField("extension").as[ContextExtension]
+      validationSettings <- cursor.downField("validationSettings").as[SigmaValidationSettings]
+      costLimit <- cursor.downField("costLimit").as[Long]
+      initCost <- cursor.downField("initCost").as[Long]
+    } yield new ErgoLikeContext(lastBlockUtxoRoot, Colls.fromArray(headers.toArray), preHeader,
+      dataBoxes, boxesToSpend, spendingTransaction, selfIndex, extension, validationSettings, costLimit, initCost)
   }
 }
